Moved posts and register to routes
